* 'develop' of https://github.com/nuls-io/nuls_2.0: (49 commits)
  refactor：rpc cmd update,chainId type change to uint16
  refactor：rpc cmd update,chainId type change to uint16
  websocket
  websocket
  websocket
  modify rpc for  new structure
  websocket
  consensus commit
  websocket
  add:AliasCmd class,it is the entry of alias rpc interface add:RpcParameterNameConstant class,it contain the contant of parameter name
  fix：rpc cmd update
  feat：rpcConstant
  coin data validate
  add:add the rpc interface of alias methods
  fix：Adaptive RPC module
  create account cmd
  lombok test
  accountState balance Repository Unit test
  accountState Repository Unit test
  feat：sendEvent
  ...

+++ tools/nuls-tools/base-tools/src/main/java/io/nuls/base/data/Address.java
@@ -54,7 +54,7 @@ public class Address {
     /**
      * chain id
      */
-    private short chainId = BaseConstant.DEFAULT_CHAIN_ID;
+    private int chainId = BaseConstant.DEFAULT_CHAIN_ID;
 
     /**
      * address type
@@ -82,7 +82,7 @@ public class Address {
         }
     }
 
-    public Address(short chainId, byte addressType, byte[] hash160) {
+    public Address(int chainId, byte addressType, byte[] hash160) {
         this.chainId = chainId;
         this.addressType = addressType;
         this.hash160 = hash160;
@@ -94,7 +94,7 @@ public class Address {
     }
 
 
-    public short getChainId() {
+    public int getChainId() {
         return chainId;
     }
 
@@ -108,7 +108,7 @@ public class Address {
             throw new NulsRuntimeException(new Exception());
         }
 
-        short chainId = ByteUtils.bytesToShort(hashs);
+        int chainId = ByteUtils.bytesToShort(hashs);
         byte addressType = hashs[2];
         byte[] content = new byte[LENGTH];
         System.arraycopy(hashs, 3, content, 0, LENGTH);
@@ -119,7 +119,7 @@ public class Address {
 
     public byte[] calcAddressbytes() {
         byte[] body = new byte[ADDRESS_LENGTH];
-        System.arraycopy(ByteUtils.shortToBytes(chainId), 0, body, 0, 2);
+        System.arraycopy(ByteUtils.shortToBytes((short)chainId), 0, body, 0, 2);
         body[2] = this.addressType;
         System.arraycopy(hash160, 0, body, 3, hash160.length);
         return body;
